Implement PartitionedContainer
Stores the particle members in separate
partitions but allows for the same
access interface as Particle

WIP: PartitionedContainer with generic access
Works on GodBolt/everything in one file but leads to compiler
crash when the container is defined in the header file?

Workaround compiler crash in PartitionedContainer

The compiler crash was likely because of the lambda in operator[] of
PartitionedContainer. It seems to be related to the return type of the
lambda. Passing the return type explicitly instead of auto deducing
or using reflection fixed it?